# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GoTunnel 内网穿透服务器 API",
        "title": "GoTunnel API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:7500",
    "basePath": "/",
    "paths": {
...
        "/api/config": {
            "get": {
                "description": "返回服务器配置（敏感信息脱敏）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "配置"
                ],
                "summary": "获取配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ServerConfigResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "description": "更新服务器配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "配置"
                ],
                "summary": "更新配置",
                "parameters": [
                    {
                        "description": "配置内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateServerConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ConfigUpdateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "dto.ConfigUpdateResponse": {
            "type": "object",
            "properties": {
                "applied_runtime_fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "restart_required_fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.ServerConfigInfo": {
            "type": "object",
            "properties": {
                "bind_addr": {
                    "type": "string"
                },
                "bind_port": {
                    "type": "integer"
                },
                "client_response_timeout_sec": {
                    "type": "integer"
                },
                "heartbeat_sec": {
                    "type": "integer"
                },
                "heartbeat_timeout": {
                    "type": "integer"
                },
                "max_client_proxies": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "dto.ServerConfigPart": {
            "type": "object",
            "properties": {
                "bind_addr": {
                    "type": "string"
                },
                "bind_port": {
                    "type": "integer",
                    "maximum": 65535,
                    "minimum": 1
                },
                "client_response_timeout_sec": {
                    "type": "integer",
                    "maximum": 300,
                    "minimum": 1
                },
                "heartbeat_sec": {
                    "type": "integer",
                    "maximum": 3600,
                    "minimum": 1
                },
                "heartbeat_timeout": {
                    "type": "integer",
                    "maximum": 3600,
                    "minimum": 1
                },
                "max_client_proxies": {
                    "type": "integer",
                    "maximum": 10000,
                    "minimum": 0
                },
                "token": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "dto.ServerConfigResponse": {
            "type": "object",
            "properties": {
                "server": {
                    "$ref": "#/definitions/dto.ServerConfigInfo"
                },
                "web": {
                    "$ref": "#/definitions/dto.WebConfigInfo"
                }
            }
        },
...
        "dto.UpdateServerConfigRequest": {
            "type": "object",
            "properties": {
                "server": {
                    "$ref": "#/definitions/dto.ServerConfigPart"
                },
                "web": {
                    "$ref": "#/definitions/dto.WebConfigPart"
                }
            }
        },
...
        "dto.WebConfigInfo": {
            "type": "object",
            "properties": {
                "bind_port": {
                    "type": "integer"
                },
                "enabled": {
                    "type": "boolean"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.WebConfigPart": {
            "type": "object",
            "properties": {
                "bind_port": {
                    "type": "integer",
                    "maximum": 65535,
                    "minimum": 1
                },
                "enabled": {
                    "type": "boolean"
                },
                "password": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 6
                },
                "username": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 3
                }
            }
        },
...
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "JWT Bearer token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}